Add plenty of features
* Introducing manifest-less device source gathering! The old school
  git clone way, implemented the config way! If you don't know how
  to make a manifest, simply comment "local_manifest_url" out in
  config.sh and edit the last 4 variables accordingly. Comment out
  the non-existent ones, add as you need. This will also require
  relevant changes on sync.sh too. I'll try to make this more
  unattended in later commits.

* Introducing messages with images embedded for Telegram chat
  announcements! telegram.sh version 0.3 introduced support for
  text stylings for all message types.
  Commit: https://github.com/fabianonline/telegram.sh/commit/707335ac53778903cf030a152ff640a81815749d
  (set parse mode for all types of messages)

  Added this feature as I tested that it won't cause any problems
  with messages sent.
  Tested-chat: https://t.me/owZ_Builds/343
  Note: The first 2 builds testing this were unable to locate build
  fail image. (build.sh was sourcing config.sh again, which resets
  RELEASES_DIR variable. Hardcoded redeclaration of RELEASES_DIR
  correctly by echoing output of `cd -` for now. Future versions
  of bash/zsh might break it. This behavior will change in a future
  commit.)

+++ sync.sh
@@ -1,17 +1,24 @@
 #!/bin/bash
 echo "Sync started for ${manifest_url}/tree/${branch}"
 if [ "${jenkins}" == "true" ]; then
-    telegram -M "Sync started for [${ROM} ${ROM_VERSION}](${manifest_url}/tree/${branch}): [See Progress](${BUILD_URL}console)"
+    telegram -i ${RELEASES_DIR}/assets/sync1.png -M "Sync started for [${ROM} ${ROM_VERSION}](${manifest_url}/tree/${branch}): [See Progress](${BUILD_URL}console)"
 else
-    telegram -M "Sync started for [${ROM} ${ROM_VERSION}](${manifest_url}/tree/${branch})"
+    telegram -i ${RELEASES_DIR}/assets/sync1.png -M "Sync started for [${ROM} ${ROM_VERSION}](${manifest_url}/tree/${branch})"
 fi
 SYNC_START=$(date +"%s")
-if [ "${official}" != "true" ]; then
-    mkdir -p .repo/local_manifests
-    if [ -f .repo/local_manifests/manifest.xml ]; then
-        rm .repo/local_manifests/manifest.xml
+if [[ ! -z ${local_manifest_url} ]]; then
+    if [ "${official}" != "true" ]; then
+        mkdir -p .repo/local_manifests
+        if [ -f .repo/local_manifests/manifest.xml ]; then
+            rm .repo/local_manifests/manifest.xml
+        fi
+        wget "${local_manifest_url}" -O .repo/local_manifests/manifest.xml
     fi
-    wget "${local_manifest_url}" -O .repo/local_manifests/manifest.xml
+else
+    git clone ${devicetree} -b ${devicebranch} ${devicepath}
+    git clone ${devicetreecommon} -b ${devicebranchcommon} ${devicepathcommon}
+    git clone ${kerneltree} -b ${kernelbranch} ${kernelpath}
+    git clone ${vendortree} -b ${vendorbranch} ${vendorpath}
 fi
 cores=$(nproc --all)
 if [ "${cores}" -gt "12" ]; then
@@ -23,11 +30,11 @@ SYNC_END=$(date +"%s")
 SYNC_DIFF=$((SYNC_END - SYNC_START))
 if [ "${syncsuccessful}" == "0" ]; then
     echo "Sync completed successfully in $((SYNC_DIFF / 60)) minute(s) and $((SYNC_DIFF % 60)) seconds"
-    telegram -N -M "Sync completed successfully in $((SYNC_DIFF / 60)) minute(s) and $((SYNC_DIFF % 60)) seconds"
+    telegram -i ${RELEASES_DIR}/assets/sync3.png -N -M "Sync completed successfully in $((SYNC_DIFF / 60)) minute(s) and $((SYNC_DIFF % 60)) seconds"
     source "${my_dir}/build.sh"
 else
     echo "Sync failed in $((SYNC_DIFF / 60)) minute(s) and $((SYNC_DIFF % 60)) seconds"
-    telegram -N -M "Sync failed in $((SYNC_DIFF / 60)) minute(s) and $((SYNC_DIFF % 60)) seconds"
+    telegram -i ${RELEASES_DIR}/assets/sync2.png -N -M "Sync failed in $((SYNC_DIFF / 60)) minute(s) and $((SYNC_DIFF % 60)) seconds"
     curl --data parse_mode=HTML --data chat_id=$TELEGRAM_CHAT --data sticker=CAADBQADGgEAAixuhBPbSa3YLUZ8DBYE --request POST https://api.telegram.org/bot$TELEGRAM_TOKEN/sendSticker
     exit 1
 fi
